+++ src/main/java/com/hcommerce/heecommerce/auth/constant/SessionConstant.java
@@ -1,4 +1,4 @@
-package com.hcommerce.heecommerce.user.constant;
+package com.hcommerce.heecommerce.auth.constant;
 
 public interface SessionConstant {
     String LOGIN_USER = "LOGIN_USER";
