+++ src/main/java/com/hcommerce/heecommerce/user/UserCommandMapper.java
@@ -1,9 +1,7 @@
 package com.hcommerce.heecommerce.user;
 
-import com.hcommerce.heecommerce.user.dto.request.UserSignUpRequestDto;
+import com.hcommerce.heecommerce.auth.dto.request.UserSignUpRequestDto;
 import org.apache.ibatis.annotations.Mapper;
-
-import java.sql.SQLIntegrityConstraintViolationException;
 
 @Mapper
 public interface UserCommandMapper {
